Fix second round of Copilot's suppressed review comments on PR #21
- CostChartBuilder.CacheHitRate: widen the token-count sum to long before
  dividing, so three near-int.MaxValue counts can't overflow into a wrong
  (possibly negative) denominator.
- PriceCatalogDatabase: fix the usage_ledger schema comment, which still
  said cost_confidence was written "Unknown" until "Phase 3 lands" - that
  phase is part of this PR, so the comment now describes the actual
  persisted format (an enum member name via Enum.ToString()).
- UsageLedgerEntry.CostConfidence: typed as the CostConfidence enum
  instead of string, so a typo'd/invalid value can no longer be persisted;
  string (de)serialization now happens only at the DB boundary
  (UsageLedger.RecordAsync / UsageRollupStore.ReadPendingEntries).
- GovernanceModelCards: SpendDisplay now reflects UnpricedRequests -
  CostUsd sums only priced requests, so when unpriced requests exist the
  card now shows "≥$X.XX (N unpriced)" instead of a plain dollar figure
  that would silently understate spend.

+++ src/TotallyHotArcRouter.Gui.Charts/CostChartBuilder.cs
@@ -497,7 +497,9 @@ public static class CostChartBuilder
     /// </summary>
     public static decimal CacheHitRate(int promptTokens, int cacheCreationTokens, int cacheReadTokens)
     {
-        var totalInputTokens = promptTokens + cacheCreationTokens + cacheReadTokens;
+        // Widened to long before summing: three int.MaxValue-range counts could overflow a 32-bit sum and
+        // silently wrap into a wrong (possibly negative) denominator.
+        var totalInputTokens = (long)promptTokens + cacheCreationTokens + cacheReadTokens;
         return totalInputTokens <= 0 ? 0m : Round(cacheReadTokens / (decimal)totalInputTokens * 100m, 1);
     }
